This is a reproducible problem with automap.
In startup.cs in the configure method you will find app.CreateTestDatabaseAccountsAndData. Here it takes you to StartupDATATestDatabaseExtenstion. From there it goes to CreateTestTenantAndAccountDetails where it creates the the method CreateTestTenantAndAccountDetails where it creates a TenantCreationRequestDto for a tenant and AccountCreationRequestDto for an account holder.

From there it goes to the CreateNewDataBaseAndTenantAdmin where it uses AutoMapper to convert the dto's to entities. I have removed all the db details etc and it jumps straight back to the calling method when it completes the line "var tenant = _mapper.Map<Tenant>(tenantModel);"

Why and how do I fix it.